Add Windows support
Previously this project had only been tested on Linux, thus it failed horribly on Windows.
Native-Extraction: Due to Mojang exporting way too many files from the natives .jar (they still do that with the new launcher....) I added a file filter for ".so" files. But forgot to add a ".dll" condition.
Command: I used ':' instead of File.PathSeparator, on windows that resolves to ';'. I did not wrap arguments with spaces with '"' cause on linux there are no spaces. Also Windows slashes '\' fail in the regex unit and need to be escaped.